UIApplicationExtensions
Moved UIApplication to new file
