+++ Sources/SwiftProcessing/Core/UIKitElements/Capture.swift
@@ -328,29 +328,6 @@ extension UIImage {
     }
 }
 
-///////////////////////////////////////////////////////////////////////////
-
-extension UIApplication {
-    
-    private class func topViewController(controller: UIViewController? = UIApplication.shared.keyWindow?.rootViewController) -> UIViewController? {
-        if let navigationController = controller as? UINavigationController {
-            return topViewController(controller: navigationController.visibleViewController)
-        }
-        if let tabController = controller as? UITabBarController {
-            if let selected = tabController.selectedViewController {
-                return topViewController(controller: selected)
-            }
-        }
-        if let presented = controller?.presentedViewController {
-            return topViewController(controller: presented)
-        }
-        return controller
-    }
-    
-    class var topViewController: UIViewController? { return topViewController() }
-    
-}
-
 extension Sketch{
     
     open func createCamera(_ position: String = "front", _ desiredFrameRate: Int? = nil) -> Camera{
